Added a Formula to install the EC2 API tools.
Unfortunately, dealing with them requires quite a bit of shell configuration.
I offered the best documentation of this configuration that I could, but it’s
fairly verbose and a bit repulsive… unfortunately, the other option (attempting
to preform the setup automatically) turns out to be even messier, and prone to
mistakes to boot.