Make Spidermonkey work with Couchdb
I had to remove the THREADSAFE flag to make it work.

Couchdb seems the reason to use Spidermonkey at this point. Without these
changes Couchdb would crash hard when loading the Spidermonkey dylib.

If the threadsafety is required then we should look into making a keg-only
Spidermonkey for Couchdb's personal use.

The nspr dependency isn't documented as required, and doesn't seem to make a
difference if it is removed. So I removed it.