Make framework install of Python actually usable.
No offense but the use of --with-framework-name option is totally wrong in the
Formula.

As noted in the Mac build [notes][1] --with-framework-name is *not* to pass
the SDK path, but to rename it (e.g. "AwesomePython.framework" instead of just
"Python.framework"). To build it as a Mac OS X framework you need to use the
--enable-framework flag instead.

The same with the other option -- to build Python universally. The flag
--enable-universalsdk is missing.

[1]: http://svn.python.org/projects/python/branches/release26-maint/Mac/README