New formula: docbook.
DocBook is XML. A lot of XML.
Unfortunately, to some people, it is still used by a lot of projects,
like git, to generate their documentation, and not all of these, like
everything except git, provide pre-built documentation for download.

This formula installs all currently released docbook-xml and the most
recent docbook-xsl package and stuffs them all in the Cellar. In
addition, a script is provided that registers all of these packages in
the global XML catalog, as noted in the caveats.

Signed-off-by: Adam Vandenberg <flangy@gmail.com>