New formula: bulk_extractor
bulk_extractor is a C++ program that scans a disk image (or any other file)
and extracts useful information—without parsing the file system. It is like a
combination of "strings" and "grep" with a whole bunch of useful patterns,
including domain names, email addresses, etc. bulk_extractor will also perform
a histogram analysis on the resulting extracted features.

Signed-off-by: Adam Vandenberg <flangy@gmail.com>