New Formula: MesaLib-GLw
Installation of the GLw library compiled against the OS X version of
OpenGL libraries. Version 7.2 is considered by many (e.g. Fink, Ubuntu)
to be the stable version of MesaLib. GLw is not provided with OS X.
By default, MesaLib first compiles GL(mesa) and then GLw against it.
This forces that any application compiled against GLw will also need
to be linked with GL(mesa), not the native OS X version of OpenGL,
since otherwise you get a version incompatibility error.
In order to avoid having a duplicated version of the OpenGL library,
and also to take advantage of any hardware acceleration that the OS X
OpenGL version could provide, it is best to compile the GLw library
against the provided OS X OpenGL library. For this purpose, some
modifications to the autoconf setting file and the GLw Makefile
were performed. In particular, the inclusion of the OS X OpenGL header
files (located in /usr/X11/include/GL) instead of the header files
provided with the mesalib package, and the linkage against the OS X
OpenGL libraries (located in /usr/X11/lib). The compilation of the
GL(mesa) libraries was also removed from the autoconf settings file.
The installation of the package files is performed manually since the
'make install' fails due to a bad detection of symbolic links.

Closes Homebrew/homebrew#6074.

Signed-off-by: Charlie Sharpsteen <source@sharpsteen.net>