New Formula: bonnie++.
Bonnie++ is a benchmark suite that is aimed at performing a number of
simple tests of hard drive and file system performance. Then you can
decide which test is important and decide how to compare different
systems after running it.

Bonnie++ was based on the code for Bonnie by Tim Bray. A summary of the
differences between Bonnie 1.0 and Bonnie++ 1.0 can be found at:

    http://www.coker.com.au/bonnie++/diff.html

Closes Homebrew/homebrew#8050.

Signed-off-by: Charlie Sharpsteen <source@sharpsteen.net>