docbook: keep XML catalog inside HOMEBREW_PREFIX
Rather than download a shell script and tell users to run it with
`sudo`, it is better to keep everything under Homebrew's control.
Luckily, the XML toolchain provides the environment variable
XML_CATALOG_FILES to allow files otehr than /etc/xml/catalog to be used.

So remove the shell script, and do all the work inside the formula
itself. When we're done, we just tell the user to add
HOMEBREW_PREFIX/etc/xml/catalog to their shell configuration, rather
than tell them to `sudo docbook-register`.

Rather than download the shell script, just download docbook 5.0, which
we need anyway.

This also simplifies the logic in a handful of docbook-using formula.

Signed-off-by: Jack Nagel <jacknagel@gmail.com>