gdb: apply upstream patch to fix OS X crash
There is a known issue with gdb 7.4 on OS X where calling `run` on twice
in the same debugging session will crash gdb. It's been fixed upstream
and it's targeted for 7.5. I've added a patch that fixes it in the
formula.

The bug report:
http://old.nabble.com/-Bug-gdb-13619--New%3A-Crash-when-running-binary-a-second-time-in-the-same-session-td33196464.html
The patch (linked from the bug report):
http://sourceware.org/ml/gdb-patches/2012-02/msg00098.html

Closes Homebrew/homebrew#8.

Signed-off-by: Jack Nagel <jacknagel@gmail.com>