gtkwave 3.3.36
Upgrage gtkwave to version 3.3.36 and remove the `--mandir` flag
because it uses the right directory now.  It apparently cannot
build its tcl support against the Lion tcl-8.5.  But if the user
has a newer tcl installed it could; therefore, the tcl code is
not explicitly disabled in this commit, even though it gives a
warning that it won't work:

    checking for Tcl_CreateInterp in -ltcl8.5... no
    configure: WARNING:
    Tcl support not enabled.

There are no self tests, but the `gtkwave` binary runs and brings
up its gui when compiled with clang and llvm from XCode-4.3.2.

Closes Homebrew/homebrew#12175.

Signed-off-by: Jack Nagel <jacknagel@gmail.com>