libdnet: fix deps conditional
Libdnet uses `autoreconf` and has deps on automake and libtool
when using XCode-4.3 or greater; however, the previous commit I
authored has the wrong conditional where it also checks if the
formula is building HEAD.  That's not right.  There is no HEAD,
and `autoreconf` is used for stable.

* Remove `if ARGV.build_head?`

This was tested on Lion by removing automake, autoconf, and libtool
then using `brew install libdnet` which now builds all the deps.
It was also tested on SL using the system autotools without any
ENV vars set, such as LIBTOOLIZE.  Tested with XCode-4.3.2, 4.0.2.

Fixes Homebrew/homebrew#12280.
Closes Homebrew/homebrew#12286.

Signed-off-by: Jack Nagel <jacknagel@gmail.com>