NetCDF: Disable Doxygen
Configure fails to disable make tasks when Doxygen is not installed and this
causes the build to fail. We don't install Doxygen docs anyway.

Also, move `ENV.fortran` to the top of the install block so that the build will
fail early if a fortran compiler is not installed.

Fixes Homebrew/homebrew#12406.