two-lame 0.3.13 + option to build cli
Upgrade twolame to version 0.3.13.
Remove the unrecognized `--disable-debug` flag.
Add a dependency on libsndfile if the user options it.
Add a user option called `--frontend`.

Twolame contains a well coded frontend cli named `twolame` that
will automatically be created if libsndfile is present.  Because
some users may want that so they can convert audio file types,
this commit adds that dependency as a user triggered option.
The frontend can't be turned off by any configure switches.  So
there is no code in the formula for args here.  This software
passes all 77 tests from `make check` when the frontend is built,
but that is also left out because it's not critical.

Closes Homebrew/homebrew#12549.

Signed-off-by: Jack Nagel <jacknagel@gmail.com>