imagemagick: remove X11.installed? conditionals
Unfortunately this won't work as was probably intended; in order for X11
paths to be injected into the build, an X11 dependency has to be active.
It's not enough for it to simply be installed.

It's possible for things to accidentally find X11 under stdenv, but
superenv actually removes X11 paths unless they are explicitly
requested.