msmtp: Explicitly request openssl
This is a preventative measure.

Msmtp can be built with either gnutls or openssl as its security lib.
By default, it will use gnutls if it exists on the user's system.  It
will use openssl only if either the user requests it or gnutls is
absent.  The current formula is silent about what security lib to use,
which means the brewer will get gnutls if it's present.

This is risky, because brewers have no idea that this is the case.
Suppose a brewer has gnutls brewed, and then brews msmtp with the
current official formula.  Then suppose the brewer removes gnutls later
but keeps msmtp.  Then msmtp will break, the user will have no idea why,
there probably will be a ticket, and no matter what the brewer will have
to rebrew.

It therefore seems better to explicitly ask msmtp to build against
openssl, since that isn't transient.  Msmtp will find the system's ssl.

For users who never brewed msmtp with gnutls already on their machines,
this commit changes nothing at all.

Closes Homebrew/homebrew#20756.

Signed-off-by: Adam Vandenberg <flangy@gmail.com>