glew: remove glu requirement from pc file
glew adds "Requires: glu" to glew.pc, even though on fresh OS X installs
there will not be a glu.pc anywhere on the system (but the library
itself is in the system's OpenGL framework).

The requirement will actually be satisfied if XQuartz is installed, but
it's only a coincidence. So I'm removing this, which should allow gource
to build on systems without XQuartz. *shrug*

Fixes Homebrew/homebrew#22025.