Add libftdi0
As it turns out, *nothing* in Homebrew that depends on libftdi is
compatible with libftdi 1.0.

avrdude doesn't even have a configure option for it, but will attempt to
link with it if libusb is also present.

lcdproc has a configure option, but instead of erroring out if it's not
found, it just silently compiles without it.

Rather than revert and wait (possibly forever) for these things to be
updated to the new library, I'm just adding the older version. It
installs alongside libftdi 1.0 without any conflicts.