libpng: move bottle block above options
Options are stored in the SoftwareSpec objects these days, so all specs
must be declared before any options, including the bottle spec.

This doesn't make a ton of sense because the bottle doesn't have
options, but until the bottle is made into a property of SoftwareSpec
itself (and that day is coming), this is what happens.

Fixes Homebrew/homebrew#22654.