poppler: enable splash backend by default
The splash backend is enabled by default, but it was made into an
explicit option in 42b31c4981c6c02749217dc9b7181a9f260dc823. This
resulted in it being *disabled* unless the option was passed, which
broke (at least) pdf2htmlex, which expects this functionality to be
present.

Fixes Homebrew/homebrew#26373.