gnuplot: fix AquaTerm 1.1.1 library detection
Gnuplot supports use of AquaTerm as an output device on OSX, provided
it detects an installed AquaTerm at build time. AquaTerm provides its
support through a framework named 'AquaTerm', which is installed in the
standard location '/Library/Frameworks'.

Gnuplot v4.6.3 does not detect the current version of AquaTerm, v1.1.1,
which was released in July 2012. This is due to a small change to the
standard AquaTerm installation, which is designed to encourage client
software to link to the framework using '-framework' style compiler
options, in place of the traditional library style options. Gnuplot's
trunk (v4.7) does now use this '-framework' style.

References:
https://github.com/AquaTerm/AquaTerm/blob/v1.1.1/aquaterm/ReleaseNotes#L1-11
https://github.com/AquaTerm/AquaTerm/blob/v1.1.1/aquaterm/INSTALL#L7-15

This formula change modifies Gnuplot 4.6.x's configure script to
generate '-framework' style options, thereby allowing a standard
installation of AquaTerm v1.1.1 to be detected. It also makes a minor
modification to an 'import' specification, which is probably only of
relevance on a case sensitive file system.

Gnuplot trunk now includes a '--with-aquaterm' configure option, which
is required before it will attempt to detect AquaTerm. This change adds
a 'noaquaterm' option to the formula to support this (so we default to
--with-aquaterm). The option will also be effective with Gnuplot v4.6.x.

Closes Homebrew/homebrew#25121.

Signed-off-by: Mike McQuaid <mike@mikemcquaid.com>