mpich2: devel/stable 3.1
This updates the stable and devel tracks to 3.1. In addition, now that the
default for the stable version of MPICH is to use shared libraries, there's no
need to have the --enable-shared option so it has been removed. The bug with
the clang library has also been fixed so that special code has also been
removed.

Closes Homebrew/homebrew#26866.

Signed-off-by: Brett Koonce <koonce@gmail.com>