xplanet: adding with-all dependency option
Adding a "with-all" option so that the necessary dependencies installed
with Xplanet match Xplanet's default configuration options, namely: PNM
support via netpbm, TrueType support via freetype, and PANGOFT2 support
via pango. This will not include JPLs CSPICE library since that is not
a current formula. Assuming the user has followed the instructions to
install and link CSPICE in Homebrew, this updated formula will include
support for that as well.

Closes Homebrew/homebrew#27403.

Signed-off-by: Mike McQuaid <mike@mikemcquaid.com>