globus-toolkit: set mac os deployment target
If not set explicitly, globus-toolkit compiles libraries with a flat ld
name space. This in turn causes it to pick up the system implementation
of GSS when linked with system framework libraries (as is the case for
the nordugrid-arc package), which causes a segfault as soon as any GSS
related functionality of globus-toolkit is used.

Closes Homebrew/homebrew#28967.

Signed-off-by: Jack Nagel <jacknagel@gmail.com>