wxmac: patch to fix 10.10 build.
We seem to have started accepting Yosemite patches, so here’s the
popular upstream proposed one that fixes 10.10 compile errors. It’s not
as thorough as a few other patches lying around, but it does the job
and is easiest to scope to just Yosemite for now without dragging
anything else in.