stunnel 5.07b2
Technically, we’re skipping a release here. 5.06 was released earlier
today but has fatal errors when built against an OpenSSL that was built
without either sslv2 or sslv3. We obviously have left sslv3 intact for
now, but you only need one of the above disabled to error out here.
This is a known error and upstream’s recommendation is for everyone
with that configuration to temporarily jump on the beta builds.