irssi: perl fix
A few things here:

* Irssi defaults to using OpenSSL now, so our ``` —enable-ssl ```
argument was getting ignored and Irssi was hooking into the system
OpenSSL regardless. I’ve changed the dep to recommended, and fixed the
configure switch. Revision to deal with the system OpenSSL tapping.

* The Perl element has gone completely rogue. See Homebrew/homebrew#34685 for discussion
& problem.

* An optional dependency on Dante so people can roll the whole SOCKS
support if they so desire. It looks for the libs Dante ships, but it
doesn’t firmly depend on them, so optional.

Closes Homebrew/homebrew#34692.

Signed-off-by: Mike McQuaid <mike@mikemcquaid.com>