libqxt: add note on upstream development cessation
Upstream aren’t going to develop this any further, and it’s likely to
break against future qt version. Have added a note of this in the
formula, and fixed some audit nits.

We could just boneyard it now, but I guess it doesn’t hurt to wait
until it actually breaks.

Closes Homebrew/homebrew#36302.

Signed-off-by: Mike McQuaid <mike@mikemcquaid.com>