parallel: remove misleading comment
`both install a 'parallel' executable. See the --without-parallel option for 'moreutils'`

That isn't actually how our conflict resolution works at this point in time. The
conflict is hard and unconditional, and isn't automatically aware of knowing which
options each formula is compiled with.

We'd certainly consider pull requests to implement a more fine-grained conflict
mechanism but presently this wording is misleading.