macvim 7.4-101
MacVim's configure already sets MACOSX_DEPLOYMENT_TARGET correctly
unless we intervene and make it incorrect:

macvim-dev/macvim@55323a5
macvim-dev/macvim#104
macvim-dev/macvim#98

Non-framework builds of Python will have PYTHONFRAMEWORKPREFIX == "" and
in that case we shouldn't add -F to LDFLAGS and CFLAGS.

OS X 10.8.4, 10.8.5, 10.9.1, and 10.9.2 have "-mno-fused-madd" in their
`python-config --cflags` but the Xcode 5.1* clang doesn't support this
option, so let's just filter it out.

Framework fix thanks to @tdsmith: "Don't try to require a framework
Python. The linkage ends up looking screwy but vim works fine even if
the Python to which it links is not a framework build."

Closes Homebrew/legacy-homebrew#50180
Closes Homebrew/legacy-homebrew#50206

Closes #60.

Signed-off-by: ilovezfs <ilovezfs@icloud.com>