nginx: stop permitting userpaths
The history on this formula is rather long but as far as I can trace userpaths
was needed because for a significant amount of time we permitted the passenger
option to pluck out the first `passenger-config` in the $PATH and use that.

That changed about two years ago now in https://github.com/Homebrew/homebrew-core/commit/7057b8baee256021b01a8a655ae45b99a2f4dc8b
and userpaths should have probably been removed at that point, but since it wasn't
explicitly breaking anything it went unnoticed and hasn't been discussed since.

Nothing about the remaining formula suggests userpaths should still be needed and
it doesn't break if you remove that element, so let's remove it.