gpx 2.4.3 (new formula)
This adds a new formula for the 3D printing utility GPX which converts
reprap gcode into Sailfish (and Makerbot) x3g.  The original author is
WHPThomas and the current maintainers are MWalker and DNewman who both
now work in the markwal fork.

This means that the letter of the two rules: canonical repo and two few
watch/star/fork, are violated, but the spirit of the rules are
satisfied. The markwal fork is the canonical fork (the debian package
comes from it), but it is relatively new so the interest measures should
use the aggregate of the three forks.

Closes #711.

Signed-off-by: Andrew Janke <andrew@apjanke.net>