perl: install libraries to sensible location
Perl's configure script, attempting to be helpful, uses shorter installation
prefixes if it detects the top-level prefix argument contains the word "Perl"
in it.

Sadly, this doesn't really work for Homebrew because whilst the installation
is actually within the Cellar containing the word "Perl" everything is symlinked
into `/usr/local` (Presuming that's where your Homebrew is), which means instead
of Perl being available in i.e. `/usr/local/lib/perl5/site_perl/5.24.0` the libraries
get symlinked into `/usr/local/lib/5.24.0` which is completely opaque to end-users
and opens up the risk of conflict with anything else which behaves in the same
way.

This tweak shoves it back into the expected sort of installation structure.